bigdata/requirements.txt
pandas
numpy
psycopg2-binary
sqlalchemy
pyarrow
requests